Added logic to handle missing serialized data (e.g. when making a run for the first time).

+++ loginformtracker.py
@@ -103,7 +103,8 @@ def dump_data(data: object, name: str) -> None:
     dump_file = str('{}.pickle').format(name)
     dump_backup_file = str('{}.pickle.bak').format(name)
 
-    shutil.copy(dump_file, dump_backup_file)
+    if os.path.exists(dump_file):
+        shutil.copy(dump_file, dump_backup_file)
 
     with open(str('{}.pickle').format(name), 'wb') as f:
         # Pickle the 'data' dictionary using the highest protocol available.
